requirements.txt
setuptools
customtkinter
psutil
selenium
bs4
sounddevice
soundfile
screen_brightness_control
matplotlib
ruaccent
TeraTTS
keyboard
pygetwindow
keras
tensorflow.keras
coqpit
flask
gradio
g2pkk
scipy
trainer
requests-oauthlib
regex
encodec
six
transformers
sympy
comtypes
gruut
aiohttp
anyascii
bangla
bnnumerizer
bnunicodenormalizer
hangul-romanize
inflect
jieba
mutagen
pypinyin
pysbd
umap
umap-learn
unidecode
jax
srt
websockets
blinker
importlib
argparse
dataclasses
faster_whisper
clearml
hangul_romanize
opencc
kaggle
cython
winshell
win10toast
face_recognition
opencv-python
wim
pywin32
sqlite3
cv2
subprocess.run
wrapt==1.11
urllib3==1.26
torch==2.3.1
torchaudio==2.3.1
numpy==1.26.4
pyaudio==0.2.13
vosk==0.3.38
gruut==2.2.3
pandas==1.4
fastapi==0.110
spacy==3.7.0
tensorflow==2.17
tensorflow-intel==2.17
tensorboard==2.17
tensorboardX==2.6.2.2
click==8.0.0
typer==0.9.4
blis==0.7.11
librosa==0.10.0